mgf.lib mgfd.lib

+++ src/framework/BackendD3D11.cpp
@@ -771,7 +771,7 @@ void BackendD3D11::GetTextSize(const wchar_t* text, mgFont* font, mgPoint* sz)
 {
 	sz->x = 0;
 	sz->y = 0;
-	int c = wcslen(text);
+	int c = (int)wcslen(text);
 	if (!c)
 		return;
 	for (int i = 0; i < c; ++i)
